requirements.txt
pyserial>=3.5
qrcode>=7.4
Pillow>=10.0
customtkinter>=5.2
pyinstaller>=6.0
